requirements.txt
asgiref==3.5.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.13
django-cors-headers==3.12.0
django-filter==21.1
django-rest-framework==0.1.0
django-stubs==1.11.0
django-stubs-ext==0.4.0
djangorestframework==3.13.1
djangorestframework-stubs==1.6.0
flake8==4.0.1
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.950
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
types-Markdown==3.3.27
types-pytz==2021.3.8
types-PyYAML==6.0.8
types-requests==2.27.30
types-urllib3==1.26.15
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
